+++ tests/Controller/UpdateEntityControllerTest.php
@@ -90,10 +90,12 @@ class UpdateEntityControllerTest extends AbstractControllerTestCase
      */
     protected function deleteEntities(): void
     {
-        foreach (self::$products as $product) {
-            $this->deleteEntity($product);
+        if (null !== self::$products) {
+            foreach (self::$products as $product) {
+                $this->deleteEntity($product);
+            }
+            self::$products = null;
         }
-        self::$products = null;
         self::$category = $this->deleteEntity(self::$category);
         self::$group = $this->deleteEntity(self::$group);
         self::$state = $this->deleteEntity(self::$state);
